move statistics to home page
