+++ apps/backoffice-v2/src/pages/Home/components/UserStatistics/UserStatistics.tsx
@@ -3,7 +3,7 @@ import { Card } from '@/common/components/atoms/Card/Card';
 import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
 import { CardContent } from '@/common/components/atoms/Card/Card.Content';
 import { CardFooter } from '@/common/components/atoms/Card/Card.Footer';
-import { useUserStatisticsLogic } from '@/pages/Statistics/components/UserStatistics/hooks/useUserStatisticsLogic/useUserStatisticsLogic';
+import { useUserStatisticsLogic } from '@/pages/Home/components/UserStatistics/hooks/useUserStatisticsLogic/useUserStatisticsLogic';
 import { TextWithNAFallback } from '@ballerine/ui';
 
 export const UserStatistics: FunctionComponent<{
