+++ apps/backoffice-v2/src/pages/Home/components/WorkflowStatistics/components/CustomLegend/CustomLegend.tsx
